// nodes/shared/types/pdf.types.ts
export interface PageMargins {
	top: number;
	bottom: number;
	left: number;
	right: number;
}